Handle correctly last net + robustness
If the netname is matching the last net in the file, the output will be incorrect, and miss the closing '))'.

This patch fixes this. It doesn't use the start of a new net as the only signal of the end of nodes in the net.
A closing parentheses of a node list, will trigger it too.

The performance should be slightly better, not that it really matters.

It also now handled netnames that are in the netlist file in double quotes, which often KiCad do.

Add a sh-bang, as well minor indent format changes.

Add small test for regexps.